Restructuring code and fixing breadcrumbs
