+++ packages/dashboard/src/components/Answers/AnswerView.tsx
@@ -1,30 +1,29 @@
 import { Grid, Typography } from "@material-ui/core"
 import React from "react"
 import { connect } from "react-redux"
+import { setQuiz } from "../../store/filter/actions"
 
 class AnswerView extends React.Component<any, any> {
   constructor(props) {
     super(props)
   }
 
-  public componentDidMount() {}
+  public componentDidMount() {
+    const answer = this.props.answers.find(
+      a => a.id === this.props.match.params.id,
+    )
+    this.props.setQuiz(answer.quizId)
+  }
 
   public render() {
     return (
       <Grid item={true} xs={12} lg={10}>
         <Grid container={true} justify="center">
-          {this.props.peerReviews.map((pr, idx) => {
-            return (
-              <Grid
-                item={true}
-                xs={12}
-                style={{ backgroundColor: "silver" }}
-                key={idx}
-              >
-                Vertaisarvio no {idx + 1}
-              </Grid>
-            )
-          })}
+          <Grid item={true} xs="auto">
+            <Typography variant="title">
+              The place for peer reviews is here
+            </Typography>
+          </Grid>
         </Grid>
       </Grid>
     )
@@ -33,11 +32,11 @@ class AnswerView extends React.Component<any, any> {
 
 const mapStateToProps = state => {
   return {
-    peerReviews: state.peerReviews,
+    answers: state.answers,
   }
 }
 
 export default connect(
   mapStateToProps,
-  null,
+  { setQuiz },
 )(AnswerView)
