+++ packages/dashboard/src/components/Answers/index.tsx
@@ -1,0 +1,4 @@
+import React from "react"
+import QuizStatistics from "./QuizStatistics"
+
+export default QuizStatistics
